Refactor guzzle based api tests to use abstract base class
